Completed all functionality of the project

+++ src/main/java/com/example/ea_java_3/domain/character/service/CharacterServiceImpl.java
@@ -1,17 +1,18 @@
 package com.example.ea_java_3.domain.character.service;
 
 import com.example.ea_java_3.domain.character.dto.CharacterMapper;
+import com.example.ea_java_3.domain.character.dto.CharacterPostDTO;
 import com.example.ea_java_3.domain.character.model.Character;
-import com.example.ea_java_3.domain.character.dto.CharacterDTO;
 import com.example.ea_java_3.domain.character.repository.CharacterRepository;
 import com.example.ea_java_3.domain.movie.model.Movie;
 import com.example.ea_java_3.domain.movie.repository.MovieRepository;
 
+import com.example.ea_java_3.exceptions.exc.CharacterInMoviesException;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * Implementation of the characterService
@@ -31,14 +32,14 @@ public class CharacterServiceImpl implements CharacterService {
     }
 
     @Override
-    public Character create(CharacterDTO dto) {
+    public Character create(CharacterPostDTO dto) {
         return characterRepo.save(mapper.toCharacter(dto));
     }
 
     @Override
     public Character getById(int id) {
         Character ch = characterRepo.findCharacterById(id);
-        Set<Movie> movies = movieRepository.findAllByCharacterId(id).stream().collect(Collectors.toSet());
+        Set<Movie> movies = new HashSet<>(movieRepository.findAllByCharacterId(id));
         ch.setMovies(movies);
         System.out.println("CHARACTER MOVIES: " + ch.getMovies());
         return ch;
@@ -46,16 +47,28 @@ public class CharacterServiceImpl implements CharacterService {
 
     @Override
     public List<Character> getAll() {
-        return characterRepo.findAll();
+        List<Character> chars = characterRepo.findAll();
+        chars.forEach(it -> {
+            Set<Movie> movies = new HashSet<>(movieRepository.findAllByCharacterId(it.getId()));
+            it.setMovies(movies);
+        });
+        return chars;
     }
 
     @Override
     public void deleteById(int id) {
+       List<Movie> charMovies =  movieRepository.findAllByCharacterId(id);
+       if (!charMovies.isEmpty()){
+           throw new CharacterInMoviesException(charMovies.size());
+       }
        characterRepo.deleteById(id);
     }
 
     @Override
-    public Character update(CharacterDTO dto) {
-        return characterRepo.save(mapper.toCharacter(dto));
+    public Character update(int id, CharacterPostDTO dto) {
+        Character ch = mapper.toCharacter(dto);
+        ch.setId(id);
+        ch.setMovies(Set.of());
+        return characterRepo.save(ch);
     }
 }
